+++ create-a-container/openapi.v1.yaml
@@ -314,13 +314,17 @@ paths:
         - { in: query, name: nodeId, schema: { type: integer }, description: Filter to a single node belonging to the site }
         - in: query
           name: user
-          schema: { type: string }
+          schema:
+            type: array
+            items: { type: string }
           description: >-
-            Owner filter. Omitted/empty returns the requesting user's own
-            containers (default). `*` returns every owner on the site for admins,
-            or, for non-admins, their own containers plus any shared with them.
-            A specific username returns that owner's containers and is admin-only
-            (except requesting yourself), yielding 403 otherwise.
+            Owner filter, sent in bracket notation (`user[0]=alice&user[1]=bob`)
+            so it always parses as a list. Omitted/empty returns everything the
+            caller may see: every container on the site for admins, or their own
+            plus any shared with them for non-admins. A list of usernames
+            returns those owners' containers for admins; for non-admins the list
+            is intersected with what they may already see (own plus shared), so
+            it can only narrow visibility.
       responses:
         '200': { description: Array of containers }
         '403': { description: 'Admin access required', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
@@ -339,7 +343,7 @@ paths:
                 customTemplate: { type: string }
                 entrypoint: { type: string }
                 nvidiaRequested: { type: boolean }
-                additionalOwners:
+                collaborators:
                   type: array
                   items: { type: string }
                   description: Usernames to share the new container with (must exist)
